+++ lib/posix/options/CMakeLists.txt
@@ -115,6 +115,7 @@ endif()
 if (NOT CONFIG_TC_PROVIDES_POSIX_TIMERS)
   zephyr_library_sources_ifdef(CONFIG_POSIX_TIMERS
     clock.c
+    clock_common.c
     timer.c
     timespec_to_timeout.c
   )
